Fix build conflicts on Mac
